// packages/clab-ui/src/core/parsing/types.ts
export interface NetemState {
  delay?: string;
  jitter?: string;
  loss?: string;
  rate?: string;
  corruption?: string;
}